requirements.txt
flask
pandas
numpy
Pillow
fastparquet
pyarrow